e2e: rename vault API compatibility test for clarity
