+++ e2e/vaultcompat/consts_test.go
@@ -1,4 +1,4 @@
-package vault
+package vaultcompat
 
 import (
 	"github.com/hashicorp/nomad/api"
