+++ e2e/vaultcompat/vault_test.go
@@ -1,4 +1,4 @@
-package vault
+package vaultcompat
 
 import (
 	"archive/zip"
